Bugfix: Incorrect AllSequentialIncreasesLessThan (nee IncreasesLessThan) Calculation
In AllSequentialIncreasesLessThan (nee IncreasesLessThan), an earlier
version of this function did not properly take into account the order
in which values were added to the moving average when calculating
the percent change. This patch fixes that error and adds an additional
criteria for the function to return `true`: there must be *at least*
`intervals` collected samples.