[Bugfix] Global adherence to max concurrent streams
If a connection in a connection pool is asked for an additional stream
that puts it over the limit of the max concurrent connections, the
connection pool will spawn a new connection. This behavior is not what
we want -- we would prefer that the new connection just wait its turn.
Setting the `http2/Transport.StrictMaxConcurrentStreams` flag will
accomplish that.